Revert "gh-12627: HnswGraphBuilder connects disconnected HNSW graph components (#13566)"
This reverts commit 217828736c41bfc68065ceb3d5b37c47116ea947.
