+++ lucene/core/src/java/org/apache/lucene/util/hnsw/HnswBuilder.java
@@ -48,5 +48,5 @@ public interface HnswBuilder {
    * components, re-ordering node ids for better delta compression) may be triggered, so callers
    * should expect this call to take some time.
    */
-  OnHeapHnswGraph getCompletedGraph() throws IOException;
+  OnHeapHnswGraph getCompletedGraph();
 }
