+++ lucene/core/src/java/org/apache/lucene/util/hnsw/HnswConcurrentMergeBuilder.java
@@ -107,17 +107,9 @@ public class HnswConcurrentMergeBuilder implements HnswBuilder {
   }
 
   @Override
-  public OnHeapHnswGraph getCompletedGraph() throws IOException {
-    if (frozen == false) {
-      // should already have been called in build(), but just in case
-      finish();
-      frozen = true;
-    }
+  public OnHeapHnswGraph getCompletedGraph() {
+    frozen = true;
     return getGraph();
-  }
-
-  private void finish() throws IOException {
-    workers[0].finish();
   }
 
   @Override
