+++ lucene/core/src/test/org/apache/lucene/search/BaseVectorSimilarityQueryTestCase.java
@@ -43,7 +43,7 @@ import org.apache.lucene.store.Directory;
 import org.apache.lucene.tests.index.RandomIndexWriter;
 import org.apache.lucene.tests.util.LuceneTestCase;
 import org.apache.lucene.tests.util.TestUtil;
-import org.apache.lucene.util.hnsw.HnswUtil;
+import org.apache.lucene.tests.util.hnsw.HnswTestUtil;
 
 @LuceneTestCase.SuppressCodecs("SimpleText")
 abstract class BaseVectorSimilarityQueryTestCase<
@@ -137,7 +137,7 @@ abstract class BaseVectorSimilarityQueryTestCase<
     try (Directory indexStore = getIndexStore(getRandomVectors(numDocs, dim));
         IndexReader reader = DirectoryReader.open(indexStore)) {
       IndexSearcher searcher = newSearcher(reader);
-      assumeTrue("graph is disconnected", HnswUtil.graphIsRooted(reader, vectorField));
+      assumeTrue("graph is disconnected", HnswTestUtil.graphIsConnected(reader, vectorField));
 
       // All vectors are above -Infinity
       Query query1 =
@@ -173,7 +173,7 @@ abstract class BaseVectorSimilarityQueryTestCase<
 
     try (Directory indexStore = getIndexStore(getRandomVectors(numDocs, dim));
         IndexReader reader = DirectoryReader.open(indexStore)) {
-      assumeTrue("graph is disconnected", HnswUtil.graphIsRooted(reader, vectorField));
+      assumeTrue("graph is disconnected", HnswTestUtil.graphIsConnected(reader, vectorField));
       IndexSearcher searcher = newSearcher(reader);
 
       Query query =
@@ -298,7 +298,7 @@ abstract class BaseVectorSimilarityQueryTestCase<
       w.commit();
 
       try (IndexReader reader = DirectoryReader.open(indexStore)) {
-        assumeTrue("graph is disconnected", HnswUtil.graphIsRooted(reader, vectorField));
+        assumeTrue("graph is disconnected", HnswTestUtil.graphIsConnected(reader, vectorField));
         IndexSearcher searcher = newSearcher(reader);
 
         Query query =
